+++ terraform/main.region.tf
@@ -113,5 +113,6 @@ module "region" {
   grafana_major_version         = var.grafana_major_version
   grafana_zone_redundancy       = var.grafana_zone_redundancy
   grafana_public_access         = var.grafana_public_access
+  grafana_private_dns_zone_ids  = var.grafana_private_dns_zone_ids
   grafana_admin_group_object_id = var.grafana_admin_group_object_id
 }